Dice Roll Scene
Added a chris folder in scenes that has all the stuff ive been working. The first being the Dice class which is a sciptable object, including an array of sides that have diffrent options for types and values. Second is the dice roller which will roll all the dice currenly in the player objects inventory in the scene you can adjust the numer of dice in it and make your own.