Built a quick and dirty die
I made some simple scriptable objects to store values and sprites in for the faces, and attached it to a DiceBody class.
The DiceBody class has an array of faces (currently set by an int var) which is filled with a small class holding the reference to the scriptable object as well as a vector which describes the normal of its face. I wrote a quick check script (which currently changes a variable for every update, probably will be changed) that can tell which side is up by just looping through the array and subtracting that normal from up. I'm currently using the magnitude but I am not sure if that will continue working, haven't tested it and might instead do an angle check later.