Player and Spell classes
Not sure exactly how some of the more complex spells are gonna be handled, set up something for the simpler ones. And I did a little work on a player class. I think an array of ints will work for mana, a collumn for each type and an int for the amount the player has.