Dice update
The ChrisDice can now have a scriptable object dice info attached to them that contains all the information of side dice and on start up will set up the look of the dice  according to specifications. The sprites are changed depending on what face is selected. Infomation is collected according to which dice sprite is facing up.