Asset Addition, FIXME
I added assets to the scene. However, I'm unsure how some of the scenes are set up to work so some things may have been set to inactive when they should be active and vice versa. The camera also hasn't been moved, which I don't know if that's how we wanna keep it.